devicetree: dt_sysinfo: make variable len a size_t
The return from strlen is size_t so make len size_t to match the
types. Also clean up some white spacing on math operators on lengths
to match fwts coding style.

Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>

+++ src/devicetree/dt_sysinfo/dt_sysinfo.c
@@ -199,21 +199,21 @@ static bool dt_fdt_stringlist_contains_last(
 	const int listlen,
 	const char *str)
 {
-	const int len = strlen(str);
+	const size_t len = strlen(str);
 	const char *p;
 
 	/* checking for either str only or last in string */
 	if (listlen < 2) /* need at least one byte plus nul */
 		return false;
 
-	p = memrchr(strlist, '\0', listlen-1);
+	p = memrchr(strlist, '\0', listlen - 1);
 	if (p) {
 		p++;
 	} else {
 		p = strlist;
 	}
 
-	return !memcmp(str, p, len+1);
+	return !memcmp(str, p, len + 1);
 }
 
 static bool machine_matches_reference_model(fwts_framework *fw,
